add router refactoring
